+++ models/serverStructure/MumieProblem.php
@@ -93,9 +93,21 @@ class MumieProblem implements \JsonSerializable
         return $this->link;
     }
 
-    public static function removeParamsFromUrl($url)
+    /**
+     * Remove language parameter from a given url.
+     *
+     * Language parameter is always the last one because it's added locally by StudIp.
+     * @param $url
+     * @return string
+     */
+    public static function removeLangParamFromUrl($url)
     {
-        return substr($url, 0, strpos($url, '?'));
+        if (strpos($url, '?lang') !== false) {
+            return substr($url, 0, strpos($url, '?lang'));
+        } else if (strpos($url, '&lang') !== false) {
+            return substr($url, 0, strpos($url, '&lang'));
+        }
+        return $url;
     }
 
     /**